+++ packages/components/src/components/input/input.stories.ts
@@ -715,6 +715,66 @@ export const Parts = {
 };
 
 /**
+ * 1. You can use the `setCustomValidity` method to set a custom validation message. This will override any native validation messages.
+ * 2. Set an empty string to clear the custom validity and make the input valid.
+ * 3. To show the validation message, call the `reportValidity` method. Originally this would show a native validation bubble, but we show the error messages inline.
+ */
+
+export const setCustomValidity = {
+  parameters: {
+    chromatic: { disableSnapshot: true }
+  },
+  render: () => {
+    return html`
+      <!-- block submit and show alert instead -->
+      <form id="validationForm" class="flex flex-col gap-2">
+        <sd-input id="custom-input" label="Input"></sd-input>
+        <div>
+          <sd-button type="submit">Submit</sd-button>
+          <sd-button id="error-button" variant="secondary">Set custom error</sd-button>
+          <sd-button id="success-button" variant="secondary">Set success</sd-button>
+          <sd-button type="reset" variant="secondary">Reset</sd-button>
+        </div>
+      </form>
+      <script type="module">
+        // Wait for custom elements to be defined
+        await Promise.all([customElements.whenDefined('sd-input'), customElements.whenDefined('sd-button')]).then(
+          () => {
+            const form = document.getElementById('validationForm');
+            const input = document.getElementById('custom-input');
+            const setErrorButton = document.getElementById('error-button');
+            const setSuccessButton = document.getElementById('success-button');
+
+            // Initial error
+            const errorMessage = \`This is an initial custom error (\${new Date().toLocaleTimeString()})\`;
+            input.setCustomValidity(errorMessage);
+            input.reportValidity();
+
+            // Show error message
+            setErrorButton.addEventListener('click', () => {
+              const errorMessage = \`This is a new custom error (\${new Date().toLocaleTimeString()})\`;
+              input.setCustomValidity(errorMessage);
+              input.reportValidity();
+            });
+
+            // Show success message
+            setSuccessButton.addEventListener('click', () => {
+              input.setCustomValidity(''); // Clear custom validity
+              input.reportValidity();
+            });
+
+            form.addEventListener('submit', event => {
+              event.preventDefault();
+              alert('All fields are valid!');
+            });
+          }
+        );
+      </script>
+    `;
+  }
+};
+
+/**
  * `sd-input` is fully accessibile via keyboard.
  */